# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Data Science Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type DatasetInfoColumnTypes = { [key: string]: string };

export type DatasetInfoPreviewItem = { [key: string]: unknown };

export type DatasetInfoMissingValues = { [key: string]: number };

export type DatasetInfoSummary = { [key: string]: unknown };

export interface DatasetInfo {
  datasetId: string;
  filename: string;
  rows: number;
  columns: number;
  columnNames: string[];
  columnTypes: DatasetInfoColumnTypes;
  preview: DatasetInfoPreviewItem[];
  missingValues: DatasetInfoMissingValues;
  summary: DatasetInfoSummary;
}

export interface CreateSessionRequest {
  datasetId: string;
  name: string;
}

export interface Session {
  id: string;
  name: string;
  datasetId: string;
  datasetFilename: string;
  datasetRows: number;
  datasetColumns: number;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export interface AgentQueryRequest {
  query: string;
}

export interface ResponseSection {
  title: string;
  content: string;
}

export type ChartDataType = (typeof ChartDataType)[keyof typeof ChartDataType];

export const ChartDataType = {
  bar: "bar",
  line: "line",
  scatter: "scatter",
  pie: "pie",
  histogram: "histogram",
  heatmap: "heatmap",
  box: "box",
} as const;

export type ChartDataDataItem = { [key: string]: unknown };

export interface ChartData {
  type: ChartDataType;
  title: string;
  data: ChartDataDataItem[];
  xKey?: string;
  yKey?: string;
  keys?: string[];
}

export interface AgentResponse {
  messageId: string;
  query: string;
  response: string;
  sections: ResponseSection[];
  charts: ChartData[];
  timestamp: string;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface Message {
  id: string;
  sessionId: string;
  role: MessageRole;
  content: string;
  sections: ResponseSection[];
  charts: ChartData[];
  createdAt: string;
}

export type UploadDatasetBody = {
  file: Blob;
};