# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Data Science Agent API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analysis
    description: Data science analysis operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /analysis/upload:
    post:
      operationId: uploadDataset
      tags: [analysis]
      summary: Upload a dataset
      description: Upload a CSV/JSON dataset for analysis
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Dataset uploaded and parsed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetInfo"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analysis/sessions:
    post:
      operationId: createSession
      tags: [analysis]
      summary: Create a new analysis session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
    get:
      operationId: listSessions
      tags: [analysis]
      summary: List all sessions
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
  /analysis/sessions/{sessionId}:
    get:
      operationId: getSession
      tags: [analysis]
      summary: Get a session by ID
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSession
      tags: [analysis]
      summary: Delete a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /analysis/sessions/{sessionId}/query:
    post:
      operationId: queryAgent
      tags: [analysis]
      summary: Ask the AI agent a data science question
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentQueryRequest"
      responses:
        "200":
          description: Agent response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analysis/sessions/{sessionId}/messages:
    get:
      operationId: getMessages
      tags: [analysis]
      summary: Get all messages in a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    DatasetInfo:
      type: object
      properties:
        datasetId:
          type: string
        filename:
          type: string
        rows:
          type: number
        columns:
          type: number
        columnNames:
          type: array
          items:
            type: string
        columnTypes:
          type: object
          additionalProperties:
            type: string
        preview:
          type: array
          items:
            type: object
            additionalProperties: true
        missingValues:
          type: object
          additionalProperties:
            type: number
        summary:
          type: object
          additionalProperties: true
      required:
        - datasetId
        - filename
        - rows
        - columns
        - columnNames
        - columnTypes
        - preview
        - missingValues
        - summary
    CreateSessionRequest:
      type: object
      properties:
        datasetId:
          type: string
        name:
          type: string
      required:
        - datasetId
        - name
    Session:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        datasetId:
          type: string
        datasetFilename:
          type: string
        datasetRows:
          type: number
        datasetColumns:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
        messageCount:
          type: number
      required:
        - id
        - name
        - datasetId
        - datasetFilename
        - datasetRows
        - datasetColumns
        - createdAt
        - updatedAt
        - messageCount
    AgentQueryRequest:
      type: object
      properties:
        query:
          type: string
      required:
        - query
    AgentResponse:
      type: object
      properties:
        messageId:
          type: string
        query:
          type: string
        response:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/ResponseSection"
        charts:
          type: array
          items:
            $ref: "#/components/schemas/ChartData"
        timestamp:
          type: string
      required:
        - messageId
        - query
        - response
        - sections
        - charts
        - timestamp
    ResponseSection:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
      required:
        - title
        - content
    ChartData:
      type: object
      properties:
        type:
          type: string
          enum: [bar, line, scatter, pie, histogram, heatmap, box]
        title:
          type: string
        data:
          type: array
          items:
            type: object
            additionalProperties: true
        xKey:
          type: string
        yKey:
          type: string
        keys:
          type: array
          items:
            type: string
      required:
        - type
        - title
        - data
    Message:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/ResponseSection"
        charts:
          type: array
          items:
            $ref: "#/components/schemas/ChartData"
        createdAt:
          type: string
      required:
        - id
        - sessionId
        - role
        - content
        - sections
        - charts
        - createdAt