# lib/api-zod/src/generated/types/agentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Data Science Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ChartData } from "./chartData";
import type { ResponseSection } from "./responseSection";

export interface AgentResponse {
  messageId: string;
  query: string;
  response: string;
  sections: ResponseSection[];
  charts: ChartData[];
  timestamp: string;
}