# lib/api-zod/src/generated/types/chartData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Data Science Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ChartDataDataItem } from "./chartDataDataItem";
import type { ChartDataType } from "./chartDataType";

export interface ChartData {
  type: ChartDataType;
  title: string;
  data: ChartDataDataItem[];
  xKey?: string;
  yKey?: string;
  keys?: string[];
}