# lib/api-zod/src/generated/types/datasetInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Data Science Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { DatasetInfoColumnTypes } from "./datasetInfoColumnTypes";
import type { DatasetInfoMissingValues } from "./datasetInfoMissingValues";
import type { DatasetInfoPreviewItem } from "./datasetInfoPreviewItem";
import type { DatasetInfoSummary } from "./datasetInfoSummary";

export interface DatasetInfo {
  datasetId: string;
  filename: string;
  rows: number;
  columns: number;
  columnNames: string[];
  columnTypes: DatasetInfoColumnTypes;
  preview: DatasetInfoPreviewItem[];
  missingValues: DatasetInfoMissingValues;
  summary: DatasetInfoSummary;
}