# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Data Science Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ChartData } from "./chartData";
import type { MessageRole } from "./messageRole";
import type { ResponseSection } from "./responseSection";

export interface Message {
  id: string;
  sessionId: string;
  role: MessageRole;
  content: string;
  sections: ResponseSection[];
  charts: ChartData[];
  createdAt: string;
}