# lib/api-zod/src/generated/types/responseSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Data Science Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ResponseSection {
  title: string;
  content: string;
}