# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Data Science Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Session {
  id: string;
  name: string;
  datasetId: string;
  datasetFilename: string;
  datasetRows: number;
  datasetColumns: number;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}